Added skipped test for generate_checksum
mock_open currently has a bug in it that prevents it from functioning
properly when dealing with byte strings.  The hashlib library requires a
byte string for calculation.  Until this is fixed in python our options are
to ignore the test, create our own mock_open based on the patch or mock the
hashlib function as well.  For the plan is to skip this new test and write a
second test that will mock the hashlib call.

Current(ish) code:
https://github.com/python/cpython/blob/master/Lib/unittest/mock.py#L2290

Bug:
https://bugs.python.org/issue23004
https://bugs.python.org/file37510/mock-open-allow-binary-data-fix-formatting.patch